+++ lib/routes.js
@@ -16,7 +16,7 @@ const _ = require('lodash'),
   files = require('./files'),
   references = require('./services/references'),
   auth = require('./auth'),
-  plugins = require('./plugins'),
+  plugins = require('./services/plugins'),
   cors = {
     origins: '*',
     methods: ['GET', 'POST', 'PUT', 'DELETE', 'OPTIONS'].join(','),
