+++ lib/services/components.test.js
@@ -10,7 +10,7 @@ const _ = require('lodash'),
   bluebird = require('bluebird'),
   upgrade = require('./upgrade'),
   schema = require('../schema'),
-  plugins = require('../plugins'),
+  plugins = require('./plugins'),
   expect = require('chai').expect,
   storage = require('../../test/fixtures/mocks/storage');
