+++ lib/services/db.test.js
@@ -3,7 +3,7 @@
 const _ = require('lodash'),
   filename = __filename.split('/').pop().split('.').shift(),
   lib = require('./' + filename),
-  plugins = require('../plugins'),
+  plugins = require('./plugins'),
   bluebird = require('bluebird'),
   expect = require('chai').expect,
   sinon = require('sinon'),
