+++ lib/services/pages.test.js
@@ -10,7 +10,7 @@ const _ = require('lodash'),
   sinon = require('sinon'),
   siteService = require('./sites'),
   timer = require('../timer'),
-  plugins = require('../plugins'),
+  plugins = require('./plugins'),
   schema = require('../schema'),
   publishService = require('./publish'),
   storage = require('../../test/fixtures/mocks/storage');
