+++ lib/services/plugins.js
