+++ lib/services/uris.test.js
@@ -6,7 +6,7 @@ const _ = require('lodash'),
   sinon = require('sinon'),
   siteService = require('./sites'),
   notifications = require('./notifications'),
-  plugins = require('../plugins'),
+  plugins = require('./plugins'),
   lib = require('./' + filename),
   storage = require('../../test/fixtures/mocks/storage');
