+++ lib/setup.js
@@ -5,7 +5,7 @@ const bluebird = require('bluebird'),
   routes = require('./routes'),
   internalBootstrap = require('./bootstrap'),
   render = require('./render'),
-  amphoraPlugins = require('./plugins'),
+  amphoraPlugins = require('./services/plugins'),
   db = require('./services/db'),
   bus = require('./services/bus');
