+++ lib/setup.test.js
@@ -6,7 +6,7 @@ const _ = require('lodash'),
   sinon = require('sinon'),
   { expect } = require('chai'),
   storage = require('../test/fixtures/mocks/storage'),
-  plugins = require('./plugins'),
+  plugins = require('./services/plugins'),
   render = require('./render'),
   bus = require('./services/bus');
